Expose enqueue APIs and fix assertionError
Exposes the core enqueue API for making it more flexible to extension. Also, fixed a subtle bug where a test would throw an exception after an assertinoError but it was only failing as the test error and not because of the assertion error. This is due to the async nature of process. Throwing from the MockWebServer would only cause an IOException and not fail the test. So, we keep it and make sure we throw it in the after method even if it were not a success.